MVP done, markup set as double

+++ src/main/java/ISell.java
@@ -1,5 +1,5 @@
 public interface ISell {
 
-    int calculateMarkup();
+    double calculateMarkup();
 
 }
